requirements.txt
# Core Libraries
requests
beautifulsoup4
html2text
tqdm
python-dotenv
nltk
numpy<2.0.0

# Embedding
sentence-transformers

# Vector DB
chromadb

# Optional UI
streamlit

# Dev Tools (Optional)
ipython
black
PyGithub
openai

#Langchain
langchain
langchain-community
langchain-openai
langchain[hub]
